Revert changes to locationReplace and redirect handling.
There doesn't seem to be a reason for these, since window.location.replace
will reload the page anyway.
locationReplace needs to hard reload the page in case of an error or a redirect.
The redirect isn't followed, so removing window.location.replace
and replacing the url in the address bar with pushState or replaceState
wouldn't be enough.